+++ monkey/__init__.py
@@ -11,6 +11,8 @@ __all__ = (
 
 _original_restx_api = None
 _injected_werkzeug_routing = False
+_original_argument_cls = None
+_original_parser_cls = None
 
 
 def get_version(pkg: str) -> typing.Union[typing.Tuple, None]:
@@ -27,17 +29,22 @@ def get_version(pkg: str) -> typing.Union[typing.Tuple, None]:
 
 
 # noinspection PyUnresolvedReferences
-def patch_restx(replace_parse_rule: bool = True, fix_restx_api: bool = True) -> None:
+def patch_restx(
+        replace_parse_rule: bool = True,
+        fix_restx_api: bool = True,
+        fix_restx_parser: bool = True,
+) -> None:
     """
     Monkey patch unmaintained `flask-restx`. This has a hidden side effects!!! See params bellow.
+
     :param replace_parse_rule: Patch werkzeug because `flask-restx` is not compatible with latest `werkzeug`
     :param fix_restx_api: fix deprecated `flask-restx.api.Api` init of `doc` endpoints after blueprint is bound
+    :param fix_restx_parser: replace failing `flask_restx.reqparse.Argument` class with fixed one
     """
-    global _original_restx_api, _injected_werkzeug_routing
-    packages = pkg_resources.working_set.by_key
-    if replace_parse_rule and "flask-restx" in packages and not _injected_werkzeug_routing and (
-            get_version("werkzeug") > (2, 1, 9) and get_version("flask-restx") <= (0, 6, 0)
-    ):
+    global _original_restx_api, _injected_werkzeug_routing, _original_argument_cls, _original_parser_cls
+
+    is_incompatible = get_version("flask") >= (2, 2, 0) and get_version("flask-restx") < (0, 6, 0)
+    if replace_parse_rule and is_incompatible and not _injected_werkzeug_routing:
         import werkzeug
         import werkzeug.routing
         from . import werkzeug_routing
@@ -48,11 +55,17 @@ def patch_restx(replace_parse_rule: bool = True, fix_restx_api: bool = True) -> 
         from werkzeug.routing import parse_rule as test_rule
         _ = test_rule
 
-    if fix_restx_api and "flask" in packages and _original_restx_api is None and (
-            get_version("flask") >= (2, 2, 0) and get_version("flask-restx") < (0, 6, 0)
-    ):
+    if fix_restx_api and _original_restx_api is None and is_incompatible:
         from . import restx_api
         import flask_restx
         _original_restx_api = flask_restx.api.Api
         flask_restx.api.Api = restx_api.ApiWrapper
         flask_restx.Api = restx_api.ApiWrapper
+
+    if fix_restx_parser and is_incompatible and _original_argument_cls is None and _original_parser_cls is None:
+        from . import restx_reqparser
+        import flask_restx
+        _original_argument_cls = flask_restx.reqparse.Argument
+        _original_parser_cls = flask_restx.reqparse.RequestParser
+        flask_restx.reqparse.Argument = restx_reqparser.Argument
+        flask_restx.reqparse.RequestParser = restx_reqparser.RequestParser